refactor(arch): replace private _temporal/_dapr imports with public re-exports [BLDX-1030]
Public SDK modules were importing directly from private _temporal and _dapr
subpackages, coupling them to specific execution/infrastructure backends.

- Re-export build_output_path and get_object_store_prefix from execution/__init__
- Re-export AsyncDaprClient from infrastructure/__init__
- Update 5 import sites across templates, credentials, clients, and app/base
- Update test mock paths to use new public import paths
